separate operations with 'AST' and 'Token' to another files
